+++ docs/shaver.md
@@ -5,4 +5,5 @@
     data-prjsf-file-name="jupyak_config.toml"
     data-prjsf-schema="./_static/jupyak-v0.schema.json"
     data-prjsf-data-format="toml"
+    data-prjsf-clear-empty="true"
 ></div>
