+++ js/src/render.tsx
@@ -4,7 +4,9 @@ import { Fragment, render } from 'react-dom';
 import { useState } from 'react';
 import Form from '@rjsf/bootstrap-4';
 
-import { TDataSet, ALL_KEYS, TUrlKey, DEFAULTS } from './tokens';
+import { TDataSet } from './tokens';
+
+import { fetchData, getFileContent, getDataSet } from './utils';
 
 export async function makeOneForm(container: HTMLElement): Promise<void> {
   const dataset = getDataSet(container);
@@ -17,14 +19,6 @@ export async function makeOneForm(container: HTMLElement): Promise<void> {
 
   const form = formComponent(dataset, { schema, formData, uiSchema });
   render(form, container);
-}
-
-export function getDataSet(el: HTMLElement): TDataSet {
-  const dataset: TDataSet = {};
-  for (const k of [...ALL_KEYS]) {
-    dataset[k] = el.dataset[k] || DEFAULTS[k];
-  }
-  return dataset;
 }
 
 export function formComponent(dataset: TDataSet, props: Partial<FormProps>) {
@@ -70,7 +64,12 @@ export function formComponent(dataset: TDataSet, props: Partial<FormProps>) {
           </li>
           <li class="list-group-item">
             <label class="form-label">Preview</label>
-            <textarea className="form-control" value={value} rows={10}></textarea>
+            <textarea
+              className="form-control"
+              value={value}
+              spellCheck={false}
+              rows={10}
+            ></textarea>
           </li>
           <li class="list-group-item">
             <label class="form-label">
@@ -102,52 +101,4 @@ export function formComponent(dataset: TDataSet, props: Partial<FormProps>) {
   };
 
   return <PRJSF />;
-}
-
-export async function getFileContent(
-  dataset: TDataSet,
-  formData: any,
-): Promise<string> {
-  const format = dataset.prjsfDataFormat;
-  switch (format) {
-    case 'json':
-      return JSON.stringify(formData, null, 2);
-    case 'toml':
-      let toml = await import('smol-toml');
-      return toml.stringify(formData);
-    case 'yaml':
-      let yaml = await import('yaml');
-      return yaml.stringify(formData);
-  }
-  return '';
-}
-
-async function fetchData(dataset: TDataSet, key: TUrlKey): Promise<any> {
-  let url = dataset[key];
-
-  if (!url) {
-    return {};
-  }
-
-  const response = await fetch(url);
-  let data: any = null;
-  const format = dataset[`${key}Format`] || 'json';
-
-  if (response.ok) {
-    switch (format) {
-      case 'json':
-        data = await response.json();
-        break;
-      case 'toml':
-        let toml = await import('smol-toml');
-        data = toml.parse(await response.text());
-        break;
-      case 'yaml':
-        let yaml = await import('yaml');
-        data = yaml.parse(await response.text());
-        break;
-    }
-  }
-
-  return data;
 }
